Updated create, delete and update methods for clubs.

+++ server/controllers/clubController.js
@@ -78,39 +78,46 @@ const createClub = async (req, res) => {
     facebook,
     } = req.body;
 
+    const { userId } = req.auth;
+
+    const existingClub = await Club.findOne({ owner: userId });
+    if (existingClub) {
+        return res.status(400).json({ error: 'Can not own more than 1 club.' });
+    }
+
     let club;
 
     try {
-    const logoBuffer = req.file ? req.file.buffer.toString('base64') : null;
-    const extension = req.file ? `image/${req.file.originalname.split('.').pop()}` : null;
+        const logoBuffer = req.file ? req.file.buffer.toString('base64') : null;
+        const extension = req.file ? `image/${req.file.originalname.split('.').pop()}` : null;
 
-    club = await Club.create({
-        name,
-        overview,
-        logo: {
-        data: logoBuffer,
-        extension: extension
-        },
-        description,
-        genre,
-        colorTheme,
-        location,
-        cost,
-        meetingsFrequency,
-        email,
-        instagram,
-        discord,
-        facebook,
-    });
+        club = await Club.create({
+            name,
+            overview,
+            logo: {
+            data: logoBuffer,
+            extension: extension
+            },
+            description,
+            genre,
+            colorTheme,
+            location,
+            cost,
+            meetingsFrequency,
+            email,
+            instagram,
+            discord,
+            facebook,
+            owner: userId
+        });
 
-    res.status(201).json(club);
+        res.status(201).json(club);
     } catch (error) {
         res.status(400).json({ error: error });
     }
 };
 
 const deleteClub = async (req, res) => {
-    //fetchs a single club based on id
     const { id } = req.params;
 
     // if club id is invalid
@@ -119,12 +126,20 @@ const deleteClub = async (req, res) => {
     }
 
     // find the club and delete it by id
-    const club = await Club.findByIdAndDelete(id);
+    const club = await Club.findById(id);
 
     // if club id does not exist
     if (!club) {
         return res.status(404).json({ error: "Club not found." });
     }
+    
+    // if user is not the owner of the club
+    if (club.owner !== req.auth.userId) {
+        return res.status(403).json({ error: "Can not delete a club you do not own." });
+    }
+
+    // Delete the club
+    await club.deleteOne();
 
     // delete all events in the club
     await Event.deleteMany({ _id: { $in: club.events } })
@@ -133,7 +148,6 @@ const deleteClub = async (req, res) => {
 };
 
 const updateClub = async (req, res) => {
-    // fetchs a single club based on id
     const { id } = req.params;
 
     // if club id is invalid
@@ -142,15 +156,21 @@ const updateClub = async (req, res) => {
     }
 
     try {
-        
-        const club = await Club.findByIdAndUpdate(id, {...req.body}, { new: true })
+        const club = await Club.findById(id)
 
-    // if club id does not exist
-    if (!club) {
-        return res.status(404).json({ error: "Club not found." });
-    }
+        // if club id does not exist
+        if (!club) {
+            return res.status(404).json({ error: "Club not found." });
+        }
 
-    res.status(200).json(club);
+        if (club.owner !== req.auth.userId) {
+            return res.status(403).json({ error: "Can not update a club you do not own." });
+        }
+
+        Object.assign(club, req.body);
+        await club.save();
+
+        res.status(200).json(club);
 
     } catch (error) {
         if (error instanceof mongoose.Error.ValidationError) {
